refactor: 抽离复用的移动端长提示 tooltip 和滚动关闭逻辑
1.  将多处重复的 useCloseOnScroll 和移动端长按压 tooltip 逻辑抽离为独立 hooks
2.  重构装备卡片、武器卡片、地下城卡片中的 tooltip 相关代码
3.  优化响应式网格布局的类名写法，替换内联样式
4.  简化元数据生成脚本的条件判断逻辑
